Migrate execute_single to StrategyOutput, implement numpy broadcasting, extract shape utilities
Migrate execute_single to return StrategyOutput with immutable
TensorStore access, completing the strategy result type migration.
All four strategy modules now return outputs instead of mutating
the cache directly. Demote store_named_outputs to test-only.

Implement full numpy-style broadcasting in broadcast_binary and
broadcast_binary_i64 with right-aligned dimension matching,
replacing the previous equal-shape and scalar-only limitation.

Extract shape utilities into onnx_shapes.rs (226 LOC):
shape_from_value_info, elem_type_from_value_info, vi_shape,
set_vi_shape, strip_symbolic_value_info, resolve_dynamic_input_shapes.
Re-exported through onnx_proto for backwards compatibility.
onnx_proto.rs reduced from 1066 to 846 LOC.
