+++ crates/dsperse/src/slicer/mod.rs
@@ -4,6 +4,7 @@ pub mod combiner;
 pub mod materializer;
 pub(crate) mod onnx_fold;
 pub mod onnx_proto;
+pub(crate) mod onnx_shapes;
 pub mod onnx_slicer;
 pub(crate) mod trace;
